gradio-demo/requirements.txt
gradio>=3.0
pillow
streamlit
